pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vla_attention"
version = "0.1.0"
description = "Cross-modal attention analysis for Vision-Language-Action models (MolmoAct / OpenVLA)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "einops>=0.7",
    "tqdm>=4.65",
]

[project.optional-dependencies]
# Heavy deps used only when running against the real model / sim on a GPU host.
runtime = [
    "torch>=2.2",
    "torchvision>=0.17",
    "transformers>=4.44",
    "accelerate>=0.30",
    "Pillow>=10.0",
]
libero = [
    "robosuite>=1.4",
]
simpler = [
    "simpler-env",
]

[tool.setuptools.packages.find]
where = ["src"]
